Implement intermediate parsing of array loops
The goal here was to create minimally optimized representation of array
indexing that is agnostic to the hardware it's on. Once we get to
`unsafe_get_collection(::CPUPointer,...)` we
1. Create an `IndexMap` for each indexing argument
2. Identify slice indexing args that map contiguous strides and combine
   them (eliminating unecessary iterations across loops and stride
   calculations)
3. Combine indexing arguments that result in dropped dimensions (e.g.,
   the last argument of `A[:,1]` drops a dimension). These are computed
   at the head of each indexing call so that they `((index - offset) *
   stride)` isn't computed with each iteration.
4. Parse into `Expr` and construct the loop

Right now this is slower than Base, requires a bunch of extr code for
parsing into `Expr` and is buggy. So it's more of concept than a full
featured implementation right now.