Clean up pure methods
Although I was able to find a couple of methods in base that were nearly
identical to these, I figured it was better to be safe. Unless I was
able to get it down to just methods from Core (i.e. not generic methods)
then I made it a generated function. I also removed pure from things
that had multiple instance of it (initializing StaticInt)