Remove unecessary code and replace @pures
This is technically a breaking change because we now assume that
dimnames is a tuple of `StaticSymbol`. However, I don't think any
downsteram libraries are using this yet. A lot of the changes that might
be "breaking" are removal of the pure methods, but the docstrings for that
weren't ever in the readme. This approach also ensures that we won't
spend so much time worrying about constant propagation with `Symbol`s.