Clean up + optimize cartesian conversion
Previously conversion from an integer to a cartesian index required a
call to `axes`, but offsets and the array size are acquired from the
array instead. This shouldn't make a difference in most cases, but if an
array type has optimized these methods over creation of an entire axis
it could be quicker.