Trying to speed up single element indexing
The approach here is typically faster than that in base without any
additional work on fusing loops across strides. This is especially
noticable when Base has to dig into parent arrays. However, single
element indexing is significantly slower here. This is due to
`to_indices`. The changes here are mostly trying to pull out as much as
possible from to_indices so that it can be inlined and avoid allocating.
It's not there yet, but it's a start.