ArrayIndex Interface
This creates a abstract type `ArrayIndex` and treats the previous
`MatrixIndex` as a constant of `ArrayIndex{2}`. This takes a slightly
different approach than what the previous `MatrixIndex` subtypes
expected. Here `ArrayIndex` is simple a type for converting a single
index and there is no guarantee that bounds checking occurs or that
it is an iterable type. Currently, this does not guarantee a specific
return type, which allows us the potential ability to return static
types if we need to chain a series of ArrayIndex types.