Initial implementation of ArrayInterfaceCore (#211)
This moves all of the non-`@require` code into an internal package. I
managed to get initial tests to run for `ArrayInterfaceCore` but there
will be several more combs through the tests to make sure we can
separate out tests that are independent of downstream packages. This
required addition of a simple proxy for `MArray` so we could test all of
the static/known traits.

Zero work has been done to actually test the ArrayInterace test at the
top level here so far, but hopefully it will be fairly straightforward
if there is an exhaustive list following `impor ArrayInterfaceCore:
...` (which should also prevent any noise downstream).