Use static indices for deletat and better param safety
Since we created our own `ArrayInterface.length` we were calling
`OneTo(static_length)` indirectly. We might as well just use the known
indices so that if the compiler ever gets smart enough it can unroll
that. We also shouldn't be doing `T.parameters[i]` as much so this fixes
some of that.