Safe functional indexing (#316)
This allows common comparison operators be parsed into indices (`<`, `<=`, `>`, `>=`). So instead of doing `x[3:end]` you can do `x[>=(3)]` (if `x` uses `ArrayInterface.getindex`. The main utility right now is that we can ensure that the indices are always inbounds using this syntax. I'm still working out the best way to communicate this info to `checkbounds` so that bounds checking is passed over (similar to how `:` becomes `Slice` which is known to be inbounds); but that will probably be more involved so I'm going to make that a separate effort.