Fully excise support for static
> If one is looking for an interface which includes functionality for statically-computed values, see StaticArrayInterface.jl.
This was separated from ArrayInterface.jl because it includes a lot of functionality that does not give substantive improvements
to the interface, and is likely to be deprecated in the near future as the compiler matures to automate a lot of its optimizations.

When digging through those functions, it's just a completely different library. It shouldn't be mixed with generic programming requirements. Those have nothing to do with generic programming, they were specifically for enforcing static guarantees, which is better done by the compiler with effects analysis these days anyways. So this entire part of the library is being scrapped.

I will open a new library, StaticArrayInterface.jl, which can be dedicated to that. But I don't want that polluting the code here, especially since a lot of it is not easy to read and not things the compiler team would recommend people doing. Along with the fact that it was the vast majority of the load time and used by almost no downstream packages.

Bye static code, good riddance. That is finally completed.